// ionicApp/src/pages/model/goat.ts
export interface GoatModel {
    id: number,
    name: string,
    price: number,
    birthdate: Date,
    breed_id: number,
    gender: string,
    localisation: string,
    description: string,
    identification: string,
    img_id: number,
    img_path: string
}

/*
export class Goat {
    id: number;
    name: string;
    price: number;
    date: any;
    breed_id: number;
    gender: string;
    localisation: string;
    description?: string;
    img_id: number;
}
*/